app/src/main/java/com/noranekoit/bajp/moe/ui/movie/MovieViewModel.kt
package com.noranekoit.bajp.moe.ui.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.noranekoit.bajp.moe.data.source.local.entity.MovieEntity
import com.noranekoit.bajp.moe.data.source.MovieRepository
import com.noranekoit.bajp.moe.vo.Resource

class MovieViewModel(private val movieRepository: MovieRepository) : ViewModel() {
    fun getMovies(): LiveData<Resource<PagedList<MovieEntity>>> =
        movieRepository.getAllMoviePopulars()
}